+++ novotest/__init__.py
@@ -1,0 +1,14 @@
+__author__ = ', '.join([
+    'Nikos Karaiskos',
+    'Mor Nitzan'
+])
+__email__ = ', '.join([
+    'nikolaos.karaiskos@mdc-berlin.de',
+    'mornitzan@fas.harvard.edu'
+])
+# __version__ = '0.2.2' # automate
+
+from . import plotting as pl
+from . import preprocessing as pp
+from . import reconstruction as rc
+from . import target_space as ts
