+++ novotest/plotting/__init__.py
@@ -1,0 +1,1 @@
+from ._plotting import plot_gene_pattern, plot_gene_patterns, plot_histogram_intestine, plot_dendrogram, plot_archetypes, plot_spatial_expression_intestine
