+++ novotest/target_space/__init__.py
@@ -1,0 +1,1 @@
+from ._target_space import construct_target_grid
